+++ src/JsonApi/Hydrator/HydratorTrait.php
@@ -13,17 +13,15 @@ use WoohooLabs\Yin\JsonApi\Schema\ResourceIdentifier;
 trait HydratorTrait
 {
     /**
-     * Determines which resource type or types can be accepted by the hydrator.
+     * Determines which resource types can be accepted by the hydrator.
      *
-     * If the hydrator can only accept one type of resources, the method should
-     * return a string. If it accepts more types, then it should return an array
-     * of strings. When such a resource is received for hydration which can't be
-     * accepted (its type doesn't match the acceptable type or types of the hydrator),
-     * a ResourceTypeUnacceptable exception will be raised.
+     * The method should return an array of acceptable resource types. When such a resource is received for hydration
+     * which can't be accepted (its type doesn't match the acceptable types of the hydrator), a ResourceTypeUnacceptable
+     * exception will be raised.
      *
-     * @return string|array
+     * @return string[]
      */
-    abstract protected function getAcceptedType();
+    abstract protected function getAcceptedTypes(): array;
 
     /**
      * Provides the attribute hydrators.
@@ -69,14 +67,10 @@ trait HydratorTrait
             throw $exceptionFactory->createResourceTypeMissingException();
         }
 
-        $acceptedType = $this->getAcceptedType();
+        $acceptedTypes = $this->getAcceptedTypes();
 
-        if (is_string($acceptedType) && $data["type"] !== $acceptedType) {
-            throw $exceptionFactory->createResourceTypeUnacceptableException($data["type"], [$acceptedType]);
-        }
-
-        if (is_array($acceptedType) && in_array($data["type"], $acceptedType, true) === false) {
-            throw $exceptionFactory->createResourceTypeUnacceptableException($data["type"], $acceptedType);
+        if (is_array($acceptedTypes) && in_array($data["type"], $acceptedTypes, true) === false) {
+            throw $exceptionFactory->createResourceTypeUnacceptableException($data["type"], $acceptedTypes);
         }
     }